server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
anyio==3.7.1